refactor: move to conftest
